Removed files, renamed files, and moved update Comic Search method to SearchFragment
