--- app/src/main/java/com/example/heroicorganizer/ApiResponse.java
@@ -1,11 +1,0 @@
-package com.example.heroicorganizer;
-
-import java.util.List;
-
-public class ApiResponse {
-    private List<Comic> results;
-
-    public List<Comic> getResults() {
-        return results;
-    }
-}
